@@ -4,75 +4,56 @@
   "description": "Angular Demo App for ngx-ui-scroll",
   "main": "./index.html",
   "scripts": {
-    "start": "npm-run-all --parallel start-api-server start-dev-server",
-    "start-dev-server": "ng serve --proxy-config proxy.conf.json",
-    "start-api-server": "node ./server/index.js",
-    "start-env": "node env",
+    "start": "ng serve --proxy-config ./demo/proxy.conf.json",
+    "start:env": "node env",
+    "start:server": "node ./server/index.js",
     "lint": "eslint . --ext .ts",
-    "test": "npm run lint && karma start",
-    "build-app": "ng build --configuration production",
-    "deploy-app": "npm run build-app && firebase deploy",
+    "test:lib": "ng test scroller",
+    "test": "npm run lint && npm run test:lib",
     "preinstall": "cd server && npm install",
-    "pack:install": "npm run build && npm pack ./dist && npm install ngx-ui-scroll-2.3.1.tgz --no-save",
-    "pack:start": "npm run pack:install && npm start",
-    "build": "node build.js",
-    "publish:lib": "npm run build && npm publish ./dist",
-    "publish:lib:next": "npm run build && npm publish --tag next ./dist"
+    "copy": "cp README.md LICENSE dist/scroller",
+    "build:lib": "ng build scroller --configuration production && npm run copy",
+    "build:lib:watch": "ng build scroller --watch",
+    "build:demo": "ng build demo --configuration production",
+    "build:demo:watch": "ng build demo --watch --configuration development",
+    "deploy:demo": "npm run build:demo && firebase deploy",
+    "pack:lib": "npm run build:lib && npm pack ./dist/scroller",
+    "publish:lib": "npm run build:lib && npm publish ./dist/scroller",
+    "publish:lib:next": "npm run build:lib && npm publish --tag next ./dist/scroller"
   },
   "license": "MIT",
   "dependencies": {
-    "@angular/animations": "^12.2.13",
-    "@angular/common": "^12.2.13",
-    "@angular/core": "^12.2.13",
-    "@angular/forms": "^12.2.13",
-    "@angular/platform-browser": "^12.2.13",
-    "@angular/platform-browser-dynamic": "^12.2.13",
-    "@angular/router": "^12.2.13",
-    "ngx-bootstrap": "^5.6.1",
-    "rxjs": "~6.6.1",
-    "tslib": "^1.13.0",
-    "vscroll": "^1.4.3",
+    "@angular/animations": "~13.3.2",
+    "@angular/common": "~13.3.2",
+    "@angular/compiler": "~13.3.2",
+    "@angular/core": "~13.3.2",
+    "@angular/forms": "~13.3.2",
+    "@angular/platform-browser": "~13.3.2",
+    "@angular/platform-browser-dynamic": "~13.3.2",
+    "@angular/router": "~13.3.2",
+    "ngx-bootstrap": "^8.0.0-RC.5",
+    "rxjs": "~7.5.1",
+    "tslib": "^2.3.1",
+    "vscroll": "^1.5.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^12.2.13",
-    "@angular/cli": "^12.2.13",
-    "@angular/compiler": "^12.2.13",
-    "@angular/compiler-cli": "^12.2.13",
-    "@rollup/plugin-node-resolve": "^7.1.3",
-    "@types/jasmine": "^3.5.0",
-    "@types/jasminewd2": "^2.0.3",
-    "@types/node": "^14.0.1",
-    "@typescript-eslint/eslint-plugin": "^4.12.0",
-    "@typescript-eslint/parser": "^4.12.0",
-    "angular2-template-loader": "^0.6.2",
-    "chalk": "^4.0.0",
-    "core-js": "~3.6.5",
-    "css-loader": "^5.0.1",
-    "dotenv": "^8.2.0",
-    "eslint": "^7.17.0",
-    "jasmine-core": "~3.7.1",
-    "karma": "^6.3.3",
-    "karma-chrome-launcher": "^3.1.0",
-    "karma-cli": "~2.0.0",
-    "karma-firefox-launcher": "^1.1.0",
-    "karma-jasmine": "^4.0.1",
-    "karma-sourcemap-loader": "^0.3.8",
-    "karma-spec-reporter": "^0.0.32",
-    "karma-webpack": "^5.0.0",
-    "npm-run-all": "^4.1.5",
-    "puppeteer": "^10.0.0",
-    "raw-loader": "^4.0.1",
-    "reflect-metadata": "^0.1.13",
-    "rollup": "^2.10.2",
-    "rollup-plugin-license": "^2.2.0",
-    "rollup-plugin-sourcemaps": "^0.6.2",
-    "sass-loader": "^8.0.2",
-    "shelljs": "^0.8.4",
-    "source-map-loader": "^0.2.4",
-    "ts-loader": "^7.0.4",
-    "typescript": "~4.2.4",
-    "uglify-js": "^3.9.3",
-    "webpack": "^5.64.0"
+    "@angular-devkit/build-angular": "~13.3.2",
+    "@angular/cli": "~13.3.2",
+    "@angular/compiler-cli": "~13.3.2",
+    "@types/jasmine": "~3.10.3",
+    "@types/node": "^17.0.8",
+    "@typescript-eslint/eslint-plugin": "^5.9.0",
+    "@typescript-eslint/parser": "^5.9.0",
+    "dotenv": "^10.0.0",
+    "eslint": "^8.6.0",
+    "jasmine-core": "~4.0.0",
+    "karma": "~6.3.10",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-firefox-launcher": "^2.1.2",
+    "karma-jasmine": "~4.0.1",
+    "ng-packagr": "^13.3.0",
+    "puppeteer": "^13.0.1",
+    "typescript": "~4.5.4"
   }
 }
