@@ -4,7 +4,7 @@ import { demos } from '../../routes';
 import { DemoSources, DemoSourceType } from '../../shared/interfaces';
 import { doLog } from '../../shared/datasource-get';
 
-import { IDatasource } from '../../../../public_api';
+import { IDatasource } from 'ngx-ui-scroll';
 
 @Component({
   selector: 'app-bidirectional-unlimited-datasource',
