@@ -3,7 +3,7 @@ import { Component } from '@angular/core';
 import { demos } from '../../routes';
 import { DemoSources, DemoSourceType } from '../../shared/interfaces';
 
-import { IDatasource } from '../../../../public_api';
+import { IDatasource } from 'ngx-ui-scroll';
 
 interface MyItem {
   index: number;
