@@ -1,9 +1,9 @@
-import { Component } from '@angular/core';
+import { Component, VERSION } from '@angular/core';
 
 import { DemoSources, DemoSourceType, MyItem } from '../shared/interfaces';
 import { globalScope as scopes, demoList as demos } from '../routes';
 
-import { Datasource } from '../../../public_api'; // from 'ngx-ui-scroll';
+import { Datasource } from 'ngx-ui-scroll';
 
 @Component({
   selector: 'app-home',
@@ -11,6 +11,7 @@ import { Datasource } from '../../../public_api'; // from 'ngx-ui-scroll';
 })
 export class HomeComponent {
 
+  angularVersion = VERSION.full;
   reloadIndex = 999;
   delay = 0;
   scopes: typeof scopes;
