@@ -7,5 +7,5 @@ import { DemoSources } from '../interfaces';
   templateUrl: './demo-sources.component.html'
 })
 export class DemoSourcesComponent {
-  @Input() sources: DemoSources;
+  @Input() sources!: DemoSources;
 }
