@@ -2,23 +2,20 @@ const path = require('path');
 const express = require('express');
 const app = express();
 
-// Heroku automagically gives us SSL
-// Lets write some middleware to redirect us
 const env = process.env.NODE_ENV || 'development';
 
-const forceSSL = (req, res, next) => {
-  if (req.headers['x-forwarded-proto'] !== 'https') {
-    return res.redirect(['https://', req.get('Host'), req.url].join(''));
-  }
-  return next();
-};
-
 if (env === 'production') {
+  const forceSSL = (req, res, next) => {
+    if (req.headers['x-forwarded-proto'] !== 'https') {
+      return res.redirect(['https://', req.get('Host'), req.url].join(''));
+    }
+    return next();
+  };
   app.use(forceSSL);
 }
 
 // Serve static files
-app.use(express.static(__dirname + '/../dist-app'));
+app.use(express.static(__dirname + '/../dist/demo'));
 
 const data = [];
 for (let i = 1; i < 1000; i++) {
@@ -51,7 +48,7 @@ app.get('/api/data', (req, res) => {
 
 // Send all requests to index.html
 app.get('/*', (req, res) =>
-  res.sendFile(path.join(__dirname + '/../dist-app/index.html'))
+  res.sendFile(path.join(__dirname + '/../dist/demo/index.html'))
 );
 // default Heroku port
 const port = process.env.PORT || 5000;
