@@ -1,11 +1,11 @@
-import { waitForAsync, ComponentFixture } from '@angular/core/testing';
+import { ComponentFixture } from '@angular/core/testing';
 import { DebugElement } from '@angular/core';
 import { By } from '@angular/platform-browser';
 
 import { Workflow, Direction, Settings } from './miscellaneous/vscroll';
 
-import { UiScrollComponent } from '../src/ui-scroll.component';
-import { IAdapter } from '../src/ui-scroll.datasource';
+import { UiScrollComponent } from '../scroller/src/ui-scroll.component';
+import { IAdapter } from '../scroller/src/ui-scroll.datasource';
 
 import { configureTestBed, configureTestBedTwo } from './scaffolding/testBed';
 import { generateDatasourceClass } from './scaffolding/datasources/class';
@@ -18,14 +18,10 @@ import { Data } from './miscellaneous/items';
 describe('Component', () => {
 
   let misc: Misc;
-  let reconfigure = true;
 
-  beforeEach(waitForAsync(() => {
-    if (!reconfigure) {
-      return;
-    }
-    reconfigure = false;
-    const fixture = configureTestBed(generateDatasourceClass('initial'), defaultTemplate);
+  beforeAll((() => {
+    const ds = generateDatasourceClass('initial');
+    const fixture = configureTestBed(ds, defaultTemplate);
     misc = new Misc(fixture);
   }));
 
