@@ -2,7 +2,7 @@ import { TestBed, ComponentFixture } from '@angular/core/testing';
 import { ComponentFixtureAutoDetect } from '@angular/core/testing';
 import { Type } from '@angular/core';
 
-import { UiScrollModule } from '../../src/ui-scroll.module';
+import { UiScrollModule } from '../../scroller/src/ui-scroll.module';
 
 import { DatasourceService } from './datasources/class';
 import {
